Move supplier assessment view from SP to shared directory
We’ll reuse it for the corresponding probation practitioner page.
