+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -62,8 +62,8 @@ import ActionPlanView from '../shared/actionPlanView'
 import SentReferral from '../../models/sentReferral'
 import ScheduleActionPlanSessionPresenter from './scheduleActionPlanSessionPresenter'
 import SupplierAssessmentDecorator from '../../decorators/supplierAssessmentDecorator'
-import SupplierAssessmentAppointmentPresenter from './supplierAssessmentAppointmentPresenter'
-import SupplierAssessmentAppointmentView from './supplierAssessmentAppointmentView'
+import SupplierAssessmentAppointmentPresenter from '../shared/supplierAssessmentAppointmentPresenter'
+import SupplierAssessmentAppointmentView from '../shared/supplierAssessmentAppointmentView'
 import SupplierAssessmentAppointmentConfirmationPresenter from './supplierAssessmentAppointmentConfirmationPresenter'
 import SupplierAssessmentAppointmentConfirmationView from './supplierAssessmentAppointmentConfirmationView'
